requirements.txt
Django>=4.2
djangorestframework>=3.14
django-filter>=24.2
gunicorn
